Some more re-structure
